rest_api/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cowsay==6.1
DateTime==5.4
dj-database-url==2.1.0
Django==5.0.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
inflect==7.0.0
Markdown==3.5.2
numpy==1.26.2
packaging==23.2
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.1
